fix: Move XCUIElement extension to shared file to avoid redeclaration
Extracted clearAndEnterText and hasFocus extensions to a shared
XCUIElementExtensions.swift file and removed duplicates from both
TrivitUITests.swift and iPadUITests.swift.
